# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SendQuote API
  description: AI-powered quoting platform for Indian businesses. Create, send, and manage professional quotes with GST compliance, e-signatures, and Razorpay payments.
  version: 2.0.0
  contact:
    name: SendQuote Support
    email: support@sendquote.in
    url: https://sendquote.in
servers:
  - url: https://sendquote.in
    description: Production server
  - url: http://localhost:3000
    description: Local development
paths:
...
  /api/quotes/send:
    post:
      operationId: sendQuote
      summary: Send a quote via email
      description: Sends a quote to the client via email using Resend. Updates quote status to 'sent' and auto-schedules AI follow-ups.
      tags:
        - Quotes
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quote_id:
                  type: string
                  format: uuid
                  description: Quote UUID
                recipient_email:
                  type: string
                  format: email
                  description: Optional override email address
      responses:
        "200":
          description: Quote sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  message:
                    type: string
                    example: "Quote sent successfully"
                  url:
                    type: string
                    format: uri
                    description: Public quote URL
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Not authorized to send this quote
        "404":
          description: Quote not found
        "409":
          description: Quote already accepted
...
  /api/clients:
    get:
      operationId: listClients
      summary: List clients
      description: Returns all clients for the authenticated user, ordered by creation date descending.
      tags:
        - CRM
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createClient
      summary: Create a client
      description: Creates a new client record for the authenticated user.
      tags:
        - CRM
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase session cookie. Set after successful login via Supabase Auth.
  schemas:
    ApiError:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: array
          description: Validation error details
          items:
            type: object
            properties:
              path:
                type: string
                description: Field path
              message:
                type: string
                description: Validation message
...
    Client:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        gst_number:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    CreateClientRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 200
        email:
          type: string
          format: email
          maxLength: 320
          nullable: true
        phone:
          type: string
          maxLength: 50
          nullable: true
        address:
          type: string
          maxLength: 1000
          nullable: true
        gst_number:
          type: string
          maxLength: 50
          nullable: true
        notes:
          type: string
          maxLength: 5000
          nullable: true
        organization_id:
          type: string
          format: uuid
          nullable: true
  responses:
    Unauthorized:
      description: Authentication required. Missing or invalid session cookie.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...